[change] move KNN from CF to helpers
