+++ tests/test_helpers.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from job.steps.collaborative_filtering.knn import KNN
+from job.helpers.knn import KNN
 from job.steps.collaborative_filtering.train_model import (
     _spotlight_transform,
     map_nearest,
